Refactor dependency graph code into its own module. Fix bugs in matching dependencies
